update param width

+++ common/ble_imu/ble_imu_types.h
@@ -47,7 +47,7 @@ typedef struct
         struct {
             uint8_t rate;
             uint8_t flags;
-            uint8_t sampling_time;
+            uint16_t sampling_time;
         } stream;
         struct {
             uint8_t rate;
